Fixed bug for vector of 8 doubles
Matlab docs indicate you have to provide not only the number in
the data type but also the size of the data type, so 8 doubles
would be fread(fid, 8, '8*double', 8), since double are 8 bytes.
This caused skipping.  By reducing this to just
fread(fid, 8, 'double'), the current number is read and no skip
occurs over valid data.  Be weary of elementPrecisionEffectiveType.
We may need to do something else there in the read() method.