Modified sample reading for matrix data
Change results in multi-element vector data being a 2D matrix
of [elements_per_sample, sample_number].  If the format is
complex, the conversion to complex is automatic.  For cases when
the format type is a matrix (3x3, for example) the output _should_
be [3, 3, num_samples] (i.e., Z-axis is sample number).  That
feature is untested.